move `_HPU_AVAILABLE`
