+++ tests/tests_pytorch/trainer/connectors/test_accelerator_connector.py
@@ -691,7 +691,7 @@ def test_unsupported_ipu_choice(mock_ipu_acc_avail, monkeypatch):
 
 @mock.patch("pytorch_lightning.accelerators.tpu._XLA_AVAILABLE", return_value=False)
 @mock.patch("pytorch_lightning.accelerators.ipu._IPU_AVAILABLE", return_value=False)
-@mock.patch("pytorch_lightning.utilities.imports._HPU_AVAILABLE", return_value=False)
+@mock.patch("pytorch_lightning.accelerators.hpu._HPU_AVAILABLE", return_value=False)
 def test_devices_auto_choice_cpu(cuda_count_0, *_):
     trainer = Trainer(accelerator="auto", devices="auto")
     assert trainer.num_devices == 1
